requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-openai
langchain-google-genai
google-generativeai
google-genai
langchain-cohere
langchain-chroma
langchain-mongodb
langchain-community
easyocr
requests
apscheduler
openai-whisper
edge-tts
pydub
streamlit-mic-recorder
ffmpeg

